attempt to use lib examples
