+++ crates/wgsparkl3d/Cargo.toml
@@ -48,5 +48,38 @@ bevy = { version = "0.15.0", features = [
 wgsparkl_testbed3d = { path = "../wgsparkl-testbed3d", features = ["dim3"] }
 obj-rs = { version = "0.7", default-features = false }
 bevy_editor_cam = "*"
-image = "0.24"
+image = "0.25"
 gltf = { version = "1.3", features = ["KHR_materials_pbrSpecularGlossiness"] }
+
+[package.metadata.docs.rs]
+cargo-args = ["-Zunstable-options", "-Zrustdoc-scrape-examples"]
+
+[[example]]
+name = "banana3"
+path = "examples/banana3.rs"
+doc-scrape-examples = true
+crate-type = ["lib"]
+
+[[example]]
+name = "elastic_cut3"
+path = "examples/elastic_cut.rs"
+doc-scrape-examples = true
+crate-type = ["lib"]
+
+[[example]]
+name = "heightfield3"
+path = "examples/heightfield3.rs"
+doc-scrape-examples = true
+crate-type = ["lib"]
+
+[[example]]
+name = "sand3"
+path = "examples/sand3.rs"
+doc-scrape-examples = true
+crate-type = ["lib"]
+
+[[example]]
+name = "default_scene"
+path = "examples/libs/default_scene.rs"
+doc-scrape-examples = true
+crate-type = ["lib"]
