+++ crates/wgsparkl3d/examples/heightfield3.rs
@@ -13,11 +13,6 @@ use wgsparkl::{
 };
 use wgsparkl_testbed3d::{AppState, PhysicsContext};
 
-#[allow(dead_code)]
-fn main() {
-    panic!("Run the `testbed3` example instead.");
-}
-
 pub fn heightfield_demo(
     mut commands: Commands,
     device: Res<RenderDevice>,
