+++ crates/wgsparkl3d/examples/testbed3.rs
@@ -1,11 +1,11 @@
 use bevy::prelude::*;
 use wgsparkl_testbed3d::{init_testbed, SceneInits};
 
-mod banana;
+mod banana3;
 mod elastic_cut3;
+mod glb_to_point_cloud_color;
 mod heightfield3;
-pub mod model_to_point_cloud;
-pub mod model_to_point_cloud_color;
+mod obj_to_point_cloud;
 mod sand3;
 
 pub fn main() {
@@ -33,15 +33,15 @@ fn register_scenes(world: &mut World) {
         ),
         (
             "elastic_model".to_string(),
-            world.register_system(model_to_point_cloud::elastic_model_demo),
+            world.register_system(obj_to_point_cloud::elastic_model_demo),
         ),
         (
             "elastic_model_colors".to_string(),
-            world.register_system(model_to_point_cloud_color::elastic_color_model_demo),
+            world.register_system(glb_to_point_cloud_color::elastic_color_model_demo),
         ),
         (
             "taichi_banana".to_string(),
-            world.register_system(banana::demo),
+            world.register_system(banana3::demo),
         ),
     ];
     let mut inits = world.resource_mut::<SceneInits>();
